@@ -1,5 +1,7 @@
 package org.httpsknighthacks.knighthacksandroid.Resources;
 
+import android.util.Log;
+
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -7,15 +9,19 @@ import java.util.Locale;
 
 public class DateTimeUtils {
 
+    public static final String DATE_TIME_STRING_PATTERN = "EE MMM dd HH:mm:ss z yyyy";
     public static final String DEFAULT_DATE_TIME_STRING_PATTERN = "yyyy-MM-dd'T'HH:mm:ss";
     public static final String YEAR_MONTH_DAY_PATTERN = "yyyyMMdd";
     public static final String TIME_STRING_PATTERN = "hh:mma";
     public static final String DATE_WEEKDAY_PATTERN = "EEEE";
+    public static final String DATE_WEEKDAY_ABBREV_PATTERN = "EE";
 
+    public static final SimpleDateFormat DATE_TIME_FORMAT = new SimpleDateFormat(DATE_TIME_STRING_PATTERN, Locale.ENGLISH);
     public static final SimpleDateFormat DEFAULT_DATE_TIME_FORMAT = new SimpleDateFormat(DEFAULT_DATE_TIME_STRING_PATTERN, Locale.ENGLISH);
     public static final SimpleDateFormat YEAR_MONTH_DAY_FORMAT = new SimpleDateFormat(YEAR_MONTH_DAY_PATTERN, Locale.ENGLISH);
     public static final SimpleDateFormat TIME_STRING_FORMAT = new SimpleDateFormat(TIME_STRING_PATTERN, Locale.ENGLISH);
     public static final SimpleDateFormat DATE_WEEKDAY_FORMAT = new SimpleDateFormat(DATE_WEEKDAY_PATTERN, Locale.ENGLISH);
+    public static final SimpleDateFormat DATE_WEEKDAY_ABBREV_FORMAT = new SimpleDateFormat(DATE_WEEKDAY_ABBREV_PATTERN, Locale.ENGLISH);
 
     public static final String DAYS_AGO = "days ago";
     public static final String MINUTES_AGO = "mins ago";
@@ -23,7 +29,7 @@ public class DateTimeUtils {
 
     public static String getWeekDayString(String dateTime) {
         try {
-            Date date = DEFAULT_DATE_TIME_FORMAT.parse(dateTime);
+            Date date = DATE_TIME_FORMAT.parse(dateTime);
 
             return DATE_WEEKDAY_FORMAT.format(date);
         } catch (ParseException ex) {
@@ -65,10 +71,10 @@ public class DateTimeUtils {
 
     public static boolean daysAreDifferent(String first, String second) {
         try {
-            Date firstDate = DEFAULT_DATE_TIME_FORMAT.parse(first);
-            Date secondDate = DEFAULT_DATE_TIME_FORMAT.parse(second);
+            Date firstDate = DATE_TIME_FORMAT.parse(first);
+            Date secondDate = DATE_TIME_FORMAT.parse(second);
 
-            return !YEAR_MONTH_DAY_FORMAT.format(firstDate).equals(YEAR_MONTH_DAY_FORMAT.format(secondDate));
+            return !DATE_WEEKDAY_ABBREV_FORMAT.format(firstDate).equals(DATE_WEEKDAY_ABBREV_FORMAT.format(secondDate));
         } catch (Exception ex) {
             return true;
         }
