@@ -3,12 +3,23 @@ package org.httpsknighthacks.knighthacksandroid.Tasks;
 
 import android.content.Context;
 import android.os.AsyncTask;
+import android.util.Log;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
 
 import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonArrayRequest;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
 
 import org.httpsknighthacks.knighthacksandroid.Models.Workshop;
 import org.httpsknighthacks.knighthacksandroid.Resources.RequestQueueSingleton;
@@ -18,61 +29,46 @@ import org.json.JSONException;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
+import java.util.List;
 
-public class WorkshopsTask extends AsyncTask<Void, Void, ArrayList<Workshop>> {
+public class WorkshopsTask {
 
-    public static final String TAG = WorkshopsTask.class.getSimpleName();
-    public static final String GET_WORKSHOPS_ROUTE = "/api/get_workshops";
+    public static final String WORKSHOPS_COLLECTION = "workshops";
 
     private WeakReference<Context> mContext;
-    private ArrayList<Workshop> mWorkshops;
     private ResponseListener<Workshop> mResponseListener;
+
+    private FirebaseFirestore mFirestore;
 
     public WorkshopsTask(Context context, ResponseListener<Workshop> responseListener) {
         this.mContext = new WeakReference<>(context);
-        this.mWorkshops = new ArrayList<>();
         this.mResponseListener = responseListener;
+        mFirestore = FirebaseFirestore.getInstance();
     }
 
-    @Override
-    protected void onPreExecute() {
-        mResponseListener.onStart();
-    }
-
-    @Override
-    protected ArrayList<Workshop> doInBackground(Void... voids) {
-        String requestURL = RequestQueueSingleton.REQUEST_API_PREFIX_URL + GET_WORKSHOPS_ROUTE;
-
-        JsonArrayRequest request = new JsonArrayRequest(Request.Method.GET, requestURL, null, new Response.Listener<JSONArray>() {
-            @Override
-            public void onResponse(JSONArray response) {
-                int numWorkshops = response.length();
-
-                for (int i = 0; i < numWorkshops; i++) {
-                    try {
-                        mWorkshops.add(new Workshop(response.getJSONObject(i)));
-                    } catch (JSONException ex) {
-                        Toast.makeText(getContext(), RequestQueueSingleton.REQUEST_ERROR_MESSAGE, Toast.LENGTH_LONG).show();
+    public void retrieveWorkshops() {
+        showLoading();
+        mFirestore.collection(WORKSHOPS_COLLECTION).get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        ArrayList<Workshop> workshops = new ArrayList<>();
+                        if (task.isSuccessful()) {
+                            workshops = new ArrayList<>();
+                            for (DocumentSnapshot document : task.getResult()) {
+                                workshops.add(document.toObject(Workshop.class));
+                            }
+                            mResponseListener.onSuccess(workshops);
+                        } else {
+                            mResponseListener.onFailure();
+                        }
+                        mResponseListener.onComplete(workshops);
                     }
-                }
-
-                mResponseListener.onSuccess(mWorkshops);
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                mResponseListener.onFailure();
-            }
-        });
-
-        RequestQueueSingleton.getInstance(getContext()).addToRequestQueue(request, TAG);
-
-        return mWorkshops;
+                });
     }
 
-    @Override
-    protected void onPostExecute(ArrayList<Workshop> workshops) {
-        mResponseListener.onComplete(workshops);
+    private void showLoading() {
+        mResponseListener.onStart();
     }
 
     public Context getContext() {
